PHASE2-Backend / api/requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.5
Django==4.2.27
django-cors-headers==4.9.0
dnspython==2.8.0
gunicorn==25.1.0
packaging==26.0
pycparser==3.0
pymongo==4.16.0
python-dotenv==1.2.2
sqlparse==0.5.5
